Update install-release.sh
"Github API is responding with a 403"

HTTP Headers, such as User-Agent, can be set in the options object. In the example below, we call the github API to find out the number of stars and forks for the request repository. This requires a custom User-Agent header as well as https.

https://github.com/request/request#custom-http-headers